Modified: backend api 변경에 따른 로직 변경
