+++ public/data/product_data.json
@@ -1,68 +1,67 @@
 [
   {
-    "id": 0,
-    "korean_name": "Tart 1",
-    "price": 1000,
-    "thumbnail_image_url": "",
-    "vegan_or_not": false,
-    "sugar_level": 1,
-    "category": { "name": "dessert", "category": "tart" },
-    "description": "",
-    "url": ""
-  },
-  {
+    "category": {
+      "category": "cake",
+      "name": "dessert"
+    },
+    "created_at": "2021-12-03T13:00:28.207Z",
     "id": 1,
-    "korean_name": "Cake 1",
-    "price": 2000,
-    "thumbnail_image_url": "",
-    "vegan_or_not": false,
-    "sugar_level": 1,
-    "category": { "name": "dessert", "category": "cake" },
-    "description": "",
-    "url": ""
+    "korean_name": "당근케이크",
+    "order_quantity": 10,
+    "price": "7000.00",
+    "thumbnail_image_url": "https://pixabay.com/get/gf740fe31fe1eee20f679fc31cacc1d29385edf12e431b4f5321556db068ae54a0865d000a5076dc98b16b213f60ed1fc.jpg",
+    "vegan_or_not": true
   },
   {
+    "category": {
+      "category": "cake",
+      "name": "dessert"
+    },
+    "created_at": "2021-12-03T13:00:28.214Z",
     "id": 2,
-    "korean_name": "Cookie 1",
-    "price": 3000,
-    "thumbnail_image_url": "",
-    "vegan_or_not": false,
-    "sugar_level": 1,
-    "category": { "name": "dessert", "category": "cookie" },
-    "description": "",
-    "url": ""
+    "korean_name": "딸기케이크",
+    "order_quantity": 5,
+    "price": "6500.00",
+    "thumbnail_image_url": "https://pixabay.com/get/g9107ddd5d067a7c77dad33fb27349d0b5046ab166bf64243357bb436707bac1af31d325116e86fc193eab9543d5f1110acf45546fa54ff82d10b37f95977b37c_1920.jpg",
+    "vegan_or_not": false
   },
   {
+    "category": {
+      "category": "cake",
+      "name": "dessert"
+    },
+    "created_at": "2021-12-03T13:00:28.218Z",
     "id": 3,
-    "korean_name": "Makaron 1",
-    "price": 4000,
-    "thumbnail_image_url": "",
-    "vegan_or_not": false,
-    "sugar_level": 1,
-    "category": { "name": "dessert", "category": "makaron" },
-    "description": "",
-    "url": ""
+    "korean_name": "치즈케이크",
+    "order_quantity": 150,
+    "price": "7000.00",
+    "thumbnail_image_url": "https://pixabay.com/photos/cheese-cake-strawberries-dessert-6263345/",
+    "vegan_or_not": true
   },
   {
+    "category": {
+      "category": "tart",
+      "name": "dessert"
+    },
+    "created_at": "2021-12-03T13:00:28.220Z",
     "id": 4,
-    "korean_name": "Cookie 2",
-    "price": 5000,
-    "thumbnail_image_url": "",
-    "vegan_or_not": false,
-    "sugar_level": 1,
-    "category": { "name": "dessert", "category": "cookie" },
-    "description": "",
-    "url": ""
+    "korean_name": "에그타르트",
+    "order_quantity": 20,
+    "price": "5000.00",
+    "thumbnail_image_url": "https://pixabay.com/get/g3f0b04163e73df082bd8d203d1433f769379d7529bfc72d9da360bee3137caddb503007c9b13d846e560a339dd8a0823fbe2a995e95d93a4258563e7154ebdbb_1920.jpg",
+    "vegan_or_not": false
   },
   {
+    "category": {
+      "category": "tart",
+      "name": "dessert"
+    },
+    "created_at": "2021-12-03T13:00:28.221Z",
     "id": 5,
-    "korean_name": "Tart 2",
-    "price": 6000,
-    "thumbnail_image_url": "",
-    "vegan_or_not": false,
-    "sugar_level": 1,
-    "category": { "name": "dessert", "category": "tart" },
-    "description": "",
-    "url": ""
+    "korean_name": "딸기타르트",
+    "order_quantity": 25,
+    "price": "5500.00",
+    "thumbnail_image_url": "https://pixabay.com/get/gf03c9e412f60a40c04fa67c330a553d8912cffe062e26c87f75a9035550dc4f31e0797e78153af34bb35311e1ade536d72c38945820541d8f932dad85d9072a7_1920.jpg",
+    "vegan_or_not": true
   }
 ]
